+++ Communication/Packets/Incoming/Avatar/SaveWardrobeOutfitEvent.cs
@@ -1,50 +1,23 @@
 ﻿using Plus.Core.FigureData;
-using Plus.Database;
 using Plus.HabboHotel.GameClients;
-
-using Dapper;
+using Plus.HabboHotel.Users.Clothing;
 
 namespace Plus.Communication.Packets.Incoming.Avatar;
 
 internal class SaveWardrobeOutfitEvent : IPacketEvent
 {
-    private readonly IFigureDataManager _figureDataManager;
-    private readonly IDatabase _database;
+    private readonly IAvatarClothingService _avatarClothingService;
 
-    public SaveWardrobeOutfitEvent(IFigureDataManager figureDataManager, IDatabase database)
+    public SaveWardrobeOutfitEvent(IAvatarClothingService avatarClothingService)
     {
-        _figureDataManager = figureDataManager;
-        _database = database;
+        _avatarClothingService = avatarClothingService;
     }
 
     public Task Parse(GameClient session, IIncomingPacket packet)
     {
-        var habbo = session.GetHabbo();
-        var clothing = habbo?.Clothing;
-        if (habbo == null || clothing == null)
-            return Task.CompletedTask;
-
         var slotId = packet.ReadInt();
         var look = packet.ReadString();
         var gender = packet.ReadString();
-        look = _figureDataManager.ProcessFigure(look, gender, clothing.GetClothingParts, true);
-
-        using (var connection = _database.Connection())
-        {
-            int rows = connection.Execute("SELECT null FROM `user_wardrobe` WHERE `user_id` = @id AND `slot_id` = @slot",
-                new { id = habbo.Id, slot = slotId });
-
-            if (rows == 1)
-            {
-                connection.Execute("UPDATE `user_wardrobe` SET `look` = @look, `gender` = @gender WHERE `user_id` = @id AND `slot_id` = @slot LIMIT 1",
-                    new { look = look, gender = gender.ToUpper(), id = habbo.Id, slot = slotId });
-            }
-            else
-            {
-                connection.Execute("INSERT INTO `user_wardrobe` (`user_id`,`slot_id`,`look`,`gender`) VALUES (@id,@slot,@look,@gender)",
-                    new { id = habbo.Id, slot = slotId, look, gender = gender.ToUpper() });
-            }
-        }
-        return Task.CompletedTask;
+        return _avatarClothingService.SaveWardrobeOutfit(session, slotId, look, gender);
     }
 }